Reintroduce QueueSizeProperty to displays
- Both ImageDisplay and PointCloud had the property previously
