+++ rviz_common/CMakeLists.txt
@@ -107,6 +107,7 @@ set(rviz_common_headers_to_moc
   include/rviz_common/properties/string_property.hpp
   include/rviz_common/properties/tf_frame_property.hpp
   include/rviz_common/properties/vector_property.hpp
+  include/rviz_common/queue_size_property.hpp
   include/rviz_common/render_panel.hpp
   src/rviz_common/scaled_image_widget.hpp
   src/rviz_common/screenshot_dialog.hpp
@@ -191,6 +192,7 @@ set(rviz_common_source_files
   src/rviz_common/properties/string_property.cpp
   src/rviz_common/properties/tf_frame_property.cpp
   src/rviz_common/properties/vector_property.cpp
+  src/rviz_common/queue_size_property.cpp
   src/rviz_common/render_panel.cpp
   src/rviz_common/ros_integration/get_topic_names_and_types.cpp
   src/rviz_common/ros_integration/init.cpp
