+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/point_cloud_display.hpp
@@ -38,6 +38,7 @@
 #include "sensor_msgs/msg/point_cloud.hpp"
 
 #include "rviz_common/ros_topic_display.hpp"
+#include "rviz_common/queue_size_property.hpp"
 #include "point_cloud_common.hpp"
 
 namespace rviz_common
@@ -66,17 +67,12 @@ namespace displays
  */
 class PointCloudDisplay : public rviz_common::RosTopicDisplay<sensor_msgs::msg::PointCloud>
 {
-  Q_OBJECT
-
 public:
   PointCloudDisplay();
 
   void reset() override;
 
   void update(float wall_dt, float ros_dt) override;
-
-private Q_SLOTS:
-  void updateQueueSize();
 
 protected:
   /** @brief Do initialization. Overridden from MessageFilterDisplay. */
@@ -85,7 +81,7 @@ protected:
   /** @brief Process a single message.  Overridden from MessageFilterDisplay. */
   void processMessage(sensor_msgs::msg::PointCloud::ConstSharedPtr cloud) override;
 
-  rviz_common::properties::IntProperty * queue_size_property_;
+  std::unique_ptr<rviz_common::QueueSizeProperty> queue_size_property_;
 
   std::unique_ptr<PointCloudCommon> point_cloud_common_;
 };
